refactor: LottoJackpotManager 에 Impl 사용 해보고 싶어서 코드 수정
