+++ src/main/java/kr/codesquad/LottoManager/LottoJackpotManager.java
@@ -6,48 +6,18 @@ import kr.codesquad.User.UserLotto;
 
 import java.util.List;
 
-public class LottoJackpotManager extends UserInput {
-    static List<Integer> JACKPOT_NUM;
-    static int BONUS_NUM;
+public interface LottoJackpotManager {
+    void setJACKPOT_NUM() throws CustomException;
 
-    public void setJACKPOT_NUM() throws CustomException {
-        this.JACKPOT_NUM = userInputJackpotNum();
-    }
+    void setBONUS_NUM() throws CustomException;
 
-    public void setBONUS_NUM() throws CustomException {
-        this.BONUS_NUM = userSetBonusNum(this.JACKPOT_NUM);
-    }
+    boolean checkNumberIsContained(List<Integer> userOneLotto, int number);
 
-    public boolean checkNumberIsContained(List<Integer> userOneLotto, int number) {
-        if (userOneLotto.contains(number)) {
-            return true;
-        } else return false;
-    }
+    int addIndexIfChecked(List<Integer> userOneLotto, int number, int index);
 
-    public int addIndexIfChecked(List<Integer> userOneLotto, int number, int index) {
-        if (checkNumberIsContained(userOneLotto, number)) {
-            return ++index;
-        } else return index;
-    }
+    void checkAllUserLottoPrize(UserLotto userLotto);
 
-    public void checkAllUserLottoPrize(UserLotto userLotto) {
-        for (EmptyLotto eachLotto : userLotto.getUserLottoList()) {
-            checkEachLottoPrize(eachLotto.getNumbers(), userLotto);
-        }
-    }
+    void checkEachLottoPrize(List<Integer> userOneLotto, UserLotto userLotto);
 
-    // prize 가 user lotto 내부에 있어서 어쩔수 없이 매개변수 2개 사용됬는데, 개선하고 싶음.
-    public void checkEachLottoPrize(List<Integer> userOneLotto, UserLotto userLotto) {
-        int index = 0;
-        for (int oneNumber : JACKPOT_NUM) {
-            index = addIndexIfChecked(userOneLotto, oneNumber, index);
-        }
-        if (index == 6) {
-            userLotto.addPrizeResultsByIndex(index - 2);
-        } else if (index == 5 && checkNumberIsContained(userOneLotto, BONUS_NUM)) {
-            userLotto.addPrizeResultsByIndex(index - 2);
-        } else if (index >= 3) {
-            userLotto.addPrizeResultsByIndex(index - 3);
-        }
-    }
+    void runJackpot(UserLotto userLotto) throws CustomException;
 }
