+++ src/main/java/kr/codesquad/Main.java
@@ -3,6 +3,7 @@ package kr.codesquad;
 
 import kr.codesquad.LottoManager.LottoGenerator;
 import kr.codesquad.LottoManager.LottoJackpotManager;
+import kr.codesquad.LottoManager.LottoJackpotManagerImpl;
 import kr.codesquad.User.UserInput;
 import kr.codesquad.User.UserLotto;
 import kr.codesquad.Utility.Utility;
@@ -15,12 +16,6 @@ public class Main {
         userLotto.printUserLotto();
     }
 
-    public static void lottoJackpotManage(UserLotto userLotto) throws CustomException {
-        LottoJackpotManager lottoJackpotManager = new LottoJackpotManager();
-        lottoJackpotManager.setJACKPOT_NUM();
-        lottoJackpotManager.setBONUS_NUM();
-        lottoJackpotManager.checkAllUserLottoPrize(userLotto);
-    }
 
     // throws 하려면 new throw 받아주는거 있어야됨. main에는 없는게 맞다.
     public static void main(String[] args) throws CustomException {
@@ -29,7 +24,9 @@ public class Main {
         UserLotto userLotto = new UserLotto(userInput.userHowManyLotto());
 
         Main.lottoGenerate(userLotto);
-        Main.lottoJackpotManage(userLotto);
+
+        LottoJackpotManager lottoJackpotManager = new LottoJackpotManagerImpl();
+        lottoJackpotManager.runJackpot(userLotto);
 
         ut.printProfit(userLotto.getPrizeResults(), userLotto.getUserAllLottoCount() * ut.getLottoPrice());
     }
